% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{Clustering_PCA}
\alias{Clustering_PCA}
\title{Clustering by PCA}
\usage{
Clustering_PCA(DATA, genes = NULL, cells = NULL, pca_max = 30)
}
\arguments{
\item{DATA}{input matrices or seurat object. Row is genes, Column is samples}

\item{genes}{vector of gene list}

\item{pca_max}{how many principal component will be output (default=30)}
}
\value{
pca object
}
\description{
Perform PCA using prcomp function
}
\examples{
Clustering_PCA(DATA, genes=Gene_list)
}
