% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{Clustering_SIMILR}
\alias{Clustering_SIMILR}
\title{Perform clustering using SIMILR}
\usage{
Clustering_SIMILR(DATA, pca_use = 1:10, seed = 1)
}
\arguments{
\item{DATA}{output from clustering_PCA}
}
\value{
similr output object
}
\description{
perform clustering using SIMILR method
}
\examples{
sim = Clustering_SIMILR(pca, pca_use = 1:10)
}
