% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{Clustering_tSNE}
\alias{Clustering_tSNE}
\title{perform tSNE clustering}
\usage{
Clustering_tSNE(DATA, pca_use = 1:10, seed = 1, perplexity = 20,
  theta = 0, max_iter = 1000)
}
\arguments{
\item{DATA}{pca result or data matrix}

\item{pca_use}{which principal component will be use for tSNE}

\item{seed}{seed}

\item{perplexity}{perplexity. this value cannot exceed (nrow(mat) -1)/3}

\item{theta}{0.0 is original tSNE. 0.5 is quick version}

\item{max_iter}{maximum number of iteration}
}
\value{
tsne result object
}
\description{
Perofrming tSNE clustering using PCA result as input
}
\examples{
tsne = Clustering_tSNE(pca, pca_use=1:10, perplexity=20)
}
