% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{Filter_PCA}
\alias{Filter_PCA}
\title{remove outlier from PCA result}
\usage{
Filter_PCA(pca, t_outlier = 20, file = NULL)
}
\arguments{
\item{pca}{pca object from Clustering_PCA}

\item{file}{if set file name, output which cells were removed.}
}
\value{
vector of cell names
}
\description{
Define outlier from PCA
}
\examples{
Filter_PCA(pca, file="output.png")
}
