% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeMatrix.R
\name{NormalizeMatrix}
\alias{NormalizeMatrix}
\title{Normalize RNA-seq matrices}
\usage{
NormalizeMatrix(mat, method = "scran", min_size = 10)
}
\arguments{
\item{mat}{single cell RNA-seq matrices}

\item{method}{method for normalize. log, quantile, rle, scran(default), cpm, tmm, tpm, upperquantile, downsample, center}

\item{min_size}{if run scran, minimum size of quick clustering. default=10}
}
\value{
normalized matrices
}
\description{
This function normalize single cell RNA-seq matrices with several different method
}
\examples{
NormalizeMatrix(mat, method="scran", min_size=10)
}
