% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{plot_PCA}
\alias{plot_PCA}
\title{plot PCA result}
\usage{
plot_PCA(pca, file = NULL, legend_file = NULL, title = NULL,
  Xcom = 1, Ycom = 2, cell_table = NULL, color_by = NULL,
  shape_by = NULL, size_by = NULL, width = 5, height = 5,
  alpha = 0.5, pallete = NULL, option = NULL)
}
\arguments{
\item{pca}{output from Clustering_PCA function}

\item{file}{graph image}

\item{legend_file}{output of legend graph}

\item{title}{title of graph}

\item{Xcom}{Specify principal component at X-axis. (Default = 1)}

\item{Ycom}{Specify principal component at Y-axis. (Default = 2)}

\item{cell_table}{data frame having cell informations. 1 of column should be "Cell". Other column could be use for specify coloring, shape, size information.}

\item{color_by}{specify color target column}

\item{shape_by}{specify shape target column}

\item{size_by}{specify size target column}

\item{width}{default = 5}

\item{height}{default = 5}

\item{pallete}{color vectors}

\item{option}{other options for ggplot drawing}
}
\value{
graph file
}
\description{
Custom PCA output using output from PCA
}
\examples{
plot_PCA(pca, file="output.png")
}
