% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{plot_SIMILR}
\alias{plot_SIMILR}
\title{make graph of SIMILR result}
\usage{
plot_SIMILR(sim, file = NULL, target = 1, legend_file = NULL,
  title = NULL, cell_table = NULL, color_by = NULL,
  shape_by = NULL, size_by = NULL, width = 5, height = 5,
  alpha = 0.5, pallete = NULL, option = NULL, le_width = NULL,
  le_height = 4)
}
\arguments{
\item{sim}{Clustering_SIMILR result object}

\item{file}{graph image}

\item{target}{which rank of SIMILR reulst will be output (default : best result)}

\item{legend_file}{output of legend graph}

\item{title}{title of graph}

\item{cell_table}{data frame having cell informations. 1 of column should be "Cell". Other column could be use for specify coloring, shape, size information.}

\item{color_by}{specify color target column}

\item{shape_by}{specify shape target column}

\item{size_by}{specify size target column}

\item{width}{default = 5}

\item{height}{default = 5}

\item{pallete}{color vectors}

\item{option}{other options for ggplot drawing}

\item{Xcom}{Specify principal component at X-axis. (Default = 1)}

\item{Ycom}{Specify principal component at Y-axis. (Default = 2)}
}
\value{
graphs
}
\description{
make graph for SIMILR results
}
\examples{
plot_SIMILR(sim, file="output.png")
}
